+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/StoredRecordChunksKafkaHandler.java
@@ -35,7 +35,6 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.NotFoundException;
-import java.util.Arrays;
 import java.util.Date;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -44,7 +43,6 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
-import static org.apache.commons.lang3.ObjectUtils.allNotNull;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_EDIFACT_RECORD_CREATED;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_MARC_BIB_FOR_ORDER_CREATED;
 import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_SRS_MARC_AUTHORITY_RECORD_CREATED;
@@ -57,6 +55,7 @@ import static org.folio.rest.jaxrs.model.Record.RecordType.MARC_AUTHORITY;
 import static org.folio.rest.jaxrs.model.Record.RecordType.MARC_BIB;
 import static org.folio.rest.jaxrs.model.Record.RecordType.MARC_HOLDING;
 import static org.folio.verticle.consumers.util.JobExecutionUtils.isNeedToSkip;
+import static org.folio.verticle.consumers.util.MarcImportEventsHandler.NO_TITLE_MESSAGE;
 
 @Component
 @Qualifier("StoredRecordChunksKafkaHandler")
@@ -219,6 +218,9 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
     for (Record record : storedRecords) {
 
       if (record.getErrorRecord() == null) {
+        String retrievedTitleFromRecord = ParsedRecordUtil.retrieveDataByField(record.getParsedRecord(), titleFieldTag, subfieldCodes);
+        if (retrievedTitleFromRecord.isEmpty()) retrievedTitleFromRecord = NO_TITLE_MESSAGE;
+
         JournalRecord journalRecord = new JournalRecord()
           .withJobExecutionId(record.getSnapshotId())
           .withSourceRecordOrder(record.getOrder())
@@ -228,8 +230,7 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
           .withActionType(CREATE)
           .withActionStatus(COMPLETED)
           .withActionDate(new Date())
-          .withTitle(allNotNull(record.getParsedRecord(), titleFieldTag)
-            ? ParsedRecordUtil.retrieveDataByField(record.getParsedRecord(), titleFieldTag, subfieldCodes) : null);
+          .withTitle(retrievedTitleFromRecord);
 
         journalRecords.add(JsonObject.mapFrom(journalRecord));
       }